+++ core/dbt/common/events/types.proto
@@ -48,17 +48,6 @@ message NodeInfo {
     NodeRelation node_relation = 10;
 }
 
-// RunResult
-message RunResultMsg {
-    string status = 1;
-    string message = 2;
-    repeated TimingInfoMsg timing_info = 3;
-    string thread = 4;
-    float execution_time = 5;
-    google.protobuf.Struct adapter_response = 6;
-    int32 num_failures = 7;
-}
-
 // GenericMessage, used for deserializing only
 message GenericMessage {
     EventInfo info = 1;
@@ -85,383 +74,6 @@ message RecordRetryException {
 message RecordRetryExceptionMsg {
     EventInfo info = 1;
     RecordRetryException data = 2;
-}
-
-// Q - Node execution
-
-// Q001
-message RunningOperationCaughtError {
-    string exc = 1;
-}
-
-message RunningOperationCaughtErrorMsg {
-    EventInfo info = 1;
-    RunningOperationCaughtError data = 2;
-}
-
-// Q002
-message CompileComplete {
-}
-
-message CompileCompleteMsg {
-    EventInfo info = 1;
-    CompileComplete data = 2;
-}
-
-// Q003
-message FreshnessCheckComplete {
-}
-
-message FreshnessCheckCompleteMsg {
-    EventInfo info = 1;
-    FreshnessCheckComplete data = 2;
-}
-
-// Q004
-message SeedHeader {
-    string header = 1;
-}
-
-message SeedHeaderMsg {
-    EventInfo info = 1;
-    SeedHeader data = 2;
-}
-
-// Skipped Q005
-
-// Q006
-message SQLRunnerException {
-    string exc = 1;
-    string exc_info = 2;
-}
-
-message SQLRunnerExceptionMsg {
-    EventInfo info = 1;
-    SQLRunnerException data = 2;
-}
-
-// Q007
-message LogTestResult {
-    NodeInfo node_info = 1;
-    string name = 2;
-    string status = 3;
-    int32 index = 4;
-    int32 num_models = 5;
-    float execution_time = 6;
-    int32 num_failures = 7;
-}
-
-message LogTestResultMsg {
-    EventInfo info = 1;
-    LogTestResult data = 2;
-}
-
-
-// Skipped Q008, Q009, Q010
-
-
-// Q011
-message LogStartLine {
-    NodeInfo node_info = 1;
-    string description = 2;
-    int32 index = 3;
-    int32 total = 4;
-}
-
-message LogStartLineMsg {
-    EventInfo info = 1;
-    LogStartLine data = 2;
-}
-
-// Q012
-message LogModelResult {
-    NodeInfo node_info = 1;
-    string description = 2;
-    string status = 3;
-    int32 index = 4;
-    int32 total = 5;
-    float execution_time = 6;
-}
-
-message LogModelResultMsg {
-    EventInfo info = 1;
-    LogModelResult data = 2;
-}
-
-// skipped Q013, Q014
-
-// Q015
-message LogSnapshotResult {
-    NodeInfo node_info = 1;
-    string description = 2;
-    string status = 3;
-    int32 index = 4;
-    int32 total = 5;
-    float execution_time = 6;
-    map<string, string> cfg = 7;
-    string result_message = 8;
-}
-
-message LogSnapshotResultMsg {
-    EventInfo info = 1;
-    LogSnapshotResult data = 2;
-}
-
-// Q016
-message LogSeedResult {
-    NodeInfo node_info = 1;
-    string status = 2;
-    string result_message = 3;
-    int32 index = 4;
-    int32 total = 5;
-    float execution_time = 6;
-    string schema = 7;
-    string relation = 8;
-}
-
-message LogSeedResultMsg {
-    EventInfo info = 1;
-    LogSeedResult data = 2;
-}
-
-// Skipped Q017
-
-// Q018
-message LogFreshnessResult {
-    string status = 1;
-    NodeInfo node_info = 2;
-    int32 index = 3;
-    int32 total = 4;
-    float execution_time = 5;
-    string source_name = 6;
-    string table_name = 7;
-}
-
-message LogFreshnessResultMsg {
-    EventInfo info = 1;
-    LogFreshnessResult data = 2;
-}
-
-
-// Skipped Q019, Q020, Q021
-
-
-// Q022
-message LogCancelLine {
-    string conn_name = 1;
-}
-
-message LogCancelLineMsg {
-    EventInfo info = 1;
-    LogCancelLine data = 2;
-}
-
-// Q023
-message DefaultSelector {
-    string name = 1;
-}
-
-message DefaultSelectorMsg {
-    EventInfo info = 1;
-    DefaultSelector data = 2;
-}
-
-// Q024
-message NodeStart {
-    NodeInfo node_info = 1;
-}
-
-message NodeStartMsg {
-    EventInfo info = 1;
-    NodeStart data = 2;
-}
-
-// Q025
-message NodeFinished {
-    NodeInfo node_info = 1;
-    RunResultMsg run_result = 2;
-}
-
-message NodeFinishedMsg {
-    EventInfo info = 1;
-    NodeFinished data = 2;
-}
-
-// Q026
-message QueryCancelationUnsupported {
-    string type = 1;
-}
-
-message QueryCancelationUnsupportedMsg {
-    EventInfo info = 1;
-    QueryCancelationUnsupported data = 2;
-}
-
-// Q027
-message ConcurrencyLine {
-    int32 num_threads = 1;
-    string target_name = 2;
-    int32 node_count = 3;
-}
-
-message ConcurrencyLineMsg {
-    EventInfo info = 1;
-    ConcurrencyLine data = 2;
-}
-
-// Q029
-message WritingInjectedSQLForNode {
-    NodeInfo node_info = 1;
-}
-
-message WritingInjectedSQLForNodeMsg {
-    EventInfo info = 1;
-    WritingInjectedSQLForNode data = 2;
-}
-
-// Q030
-message NodeCompiling {
-    NodeInfo node_info = 1;
-}
-
-message NodeCompilingMsg {
-    EventInfo info = 1;
-    NodeCompiling data = 2;
-}
-
-// Q031
-message NodeExecuting {
-    NodeInfo node_info = 1;
-}
-
-message NodeExecutingMsg {
-    EventInfo info = 1;
-    NodeExecuting data = 2;
-}
-
-// Q032
-message LogHookStartLine {
-    NodeInfo node_info = 1;
-    string statement = 2;
-    int32 index = 3;
-    int32 total = 4;
-}
-
-message LogHookStartLineMsg {
-    EventInfo info = 1;
-    LogHookStartLine data = 2;
-}
-
-// Q033
-message LogHookEndLine {
-    NodeInfo node_info = 1;
-    string statement = 2;
-    string status = 3;
-    int32 index = 4;
-    int32 total = 5;
-    float execution_time = 6;
-}
-
-message LogHookEndLineMsg {
-    EventInfo info = 1;
-    LogHookEndLine data = 2;
-}
-
-// Q034
-message SkippingDetails {
-    NodeInfo node_info = 1;
-    string resource_type = 2;
-    string schema = 3;
-    string node_name = 4;
-    int32 index = 5;
-    int32 total = 6;
-}
-
-message SkippingDetailsMsg {
-    EventInfo info = 1;
-    SkippingDetails data = 2;
-}
-
-// Q035
-message NothingToDo {
-}
-
-message NothingToDoMsg {
-    EventInfo info = 1;
-    NothingToDo data = 2;
-}
-
-// Q036
-message RunningOperationUncaughtError {
-    string exc = 1;
-}
-
-message RunningOperationUncaughtErrorMsg {
-    EventInfo info = 1;
-    RunningOperationUncaughtError data = 2;
-}
-
-// Q037
-message EndRunResult {
-    repeated RunResultMsg results = 1;
-    float elapsed_time = 2;
-    google.protobuf.Timestamp generated_at = 3;
-    bool success = 4;
-}
-
-message EndRunResultMsg {
-    EventInfo info = 1;
-    EndRunResult data = 2;
-}
-
-// Q038
-message NoNodesSelected {
-}
-
-message NoNodesSelectedMsg {
-    EventInfo info = 1;
-    NoNodesSelected data = 2;
-}
-
-// Q039
-message CommandCompleted {
-    string command = 1;
-    bool success = 2;
-    google.protobuf.Timestamp completed_at = 3;
-    float elapsed = 4;
-}
-
-message CommandCompletedMsg {
-    EventInfo info = 1;
-    CommandCompleted data = 2;
-}
-
-// Q041
-message ShowNode {
-    string node_name = 1;
-    string preview = 2;
-    bool is_inline = 3;
-    string output_format = 4;
-    string unique_id = 5;
-}
-
-message ShowNodeMsg {
-    EventInfo info = 1;
-    ShowNode data = 2;
-}
-
-// Q042
-message CompiledNode {
-    string node_name = 1;
-    string compiled = 2;
-    bool is_inline = 3;
-    string output_format = 4;
-    string unique_id = 5;
-}
-
-message CompiledNodeMsg {
-    EventInfo info = 1;
-    CompiledNode data = 2;
 }
 
 // W - Node testing
