+++ core/dbt/task/build.py
@@ -13,7 +13,7 @@ from dbt.task.test import TestSelector
 from dbt.task.base import BaseRunner
 from dbt.contracts.results import RunResult, RunStatus
 from dbt.common.events.functions import fire_event
-from dbt.common.events.types import LogStartLine, LogModelResult
+from dbt.events.types import LogStartLine, LogModelResult
 from dbt.common.events.base_types import EventLevel
