+++ core/dbt/task/run.py
@@ -33,14 +33,14 @@ from dbt.adapters.events.types import (
     FinishedRunningStats,
 )
 from dbt.common.events.functions import fire_event, get_invocation_id
-from dbt.common.events.types import (
-    Formatting,
+from dbt.common.events.types import Formatting
+from dbt.common.events.base_types import EventLevel
+from dbt.events.types import (
     LogModelResult,
     LogStartLine,
     LogHookEndLine,
     LogHookStartLine,
 )
-from dbt.common.events.base_types import EventLevel
 from dbt.logger import (
     TextOnly,
     HookMetadata,
