+++ core/dbt/task/seed.py
@@ -10,13 +10,13 @@ from dbt.common.exceptions import DbtInternalError
 from dbt.graph import ResourceTypeSelector
 from dbt.logger import TextOnly
 from dbt.common.events.functions import fire_event
-from dbt.common.events.types import (
+from dbt.common.events.types import Formatting
+from dbt.common.events.base_types import EventLevel
+from dbt.events.types import (
     SeedHeader,
-    Formatting,
     LogSeedResult,
     LogStartLine,
 )
-from dbt.common.events.base_types import EventLevel
 from dbt.node_types import NodeType
 from dbt.contracts.results import NodeStatus
