+++ core/dbt/task/show.py
@@ -7,7 +7,8 @@ from dbt.contracts.graph.nodes import SeedNode
 from dbt.contracts.results import RunResult, RunStatus
 from dbt.common.events.base_types import EventLevel
 from dbt.common.events.functions import fire_event
-from dbt.common.events.types import ShowNode, Note
+from dbt.common.events.types import Note
+from dbt.events.types import ShowNode
 from dbt.common.exceptions import DbtRuntimeError
 from dbt.task.compile import CompileTask, CompileRunner
 from dbt.task.seed import SeedRunner
