+++ core/dbt/task/sql.py
@@ -12,7 +12,7 @@ from dbt.contracts.sql import (
     ResultTable,
 )
 from dbt.common.events.functions import fire_event
-from dbt.common.events.types import SQLRunnerException
+from dbt.events.types import SQLRunnerException
 from dbt.task.compile import CompileRunner
